requirements.txt
anyio==3.7.1
attrs==23.1.0
black==23.7.0
certifi @ file:///croot/certifi_1669749558946/work/certifi
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.4
coverage==7.1.0
distlib==0.3.6
fastapi==0.89.1
filelock==3.12.2
greenlet==2.0.2
h11==0.14.0
identify==2.5.24
idna==3.4
iniconfig==2.0.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
pluggy==1.2.0
pre-commit==3.0.2
pycodestyle==2.10.0
pydantic==1.10.4
pydocstyle==6.3.0
pysqlite3==0.5.0
pytest==7.2.1
pytest-cov==4.0.0
PyYAML==6.0
requests==2.28.2
sniffio==1.3.0
snowballstemmer==2.2.0
SQLAlchemy==2.0.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.23.1
